[Refactoring] (1) Move Task to entity (2) Create Tasks to avoid primitive obsession.
