+++ src/test/java/com/codurance/training/tasks/ApplicationTest.java
@@ -6,6 +6,8 @@ import java.io.InputStreamReader;
 import java.io.PipedInputStream;
 import java.io.PipedOutputStream;
 import java.io.PrintWriter;
+
+import com.codurance.training.tasks.entity.TaskList;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
